refactor(python): nest numpy_utils under _internal/utils/
Move michelangelo.lib._internal.numpy_utils ->
michelangelo.lib._internal.utils.numpy_utils, matching the codebase's
established '<private-bucket>/utils/<name>_utils/' convention (top-level
_internal/utils/{file_utils,reflection_utils} and
model_manager/_private/utils/*) and mirroring the internal SDK's
shared/utils/numpy_utils/ layout. errors.py stays at _internal/ since it
is a shared exception, not a util.

Update the re-export __init__, intra-package imports, and the two callers
in trainer/torch. Add missing docstrings to the moved test classes and
methods so ruff check passes locally under the pinned linter.
