+++ python/michelangelo/lib/_internal/utils/numpy_utils/__init__.py
@@ -1,12 +1,12 @@
 """Numpy padding, dtype, and PyArrow conversion utilities shared across lib."""
 
-from michelangelo.lib._internal.numpy_utils.pad import pad_ragged_tensor
-from michelangelo.lib._internal.numpy_utils.pyarrow_conversion import (
+from michelangelo.lib._internal.utils.numpy_utils.pad import pad_ragged_tensor
+from michelangelo.lib._internal.utils.numpy_utils.pyarrow_conversion import (
     assemble_output_table,
     numpy_to_pyarrow,
     pyarrow_to_numpy,
 )
-from michelangelo.lib._internal.numpy_utils.sentinel import (
+from michelangelo.lib._internal.utils.numpy_utils.sentinel import (
     BOOL_SENTINEL,
     BYTES_SENTINEL,
     FLOAT_SENTINEL,
@@ -14,7 +14,7 @@ from michelangelo.lib._internal.numpy_utils.sentinel import (
     STRING_SENTINEL,
     sentinel_for_numpy_dtype,
 )
-from michelangelo.lib._internal.numpy_utils.type import infer_dtype
+from michelangelo.lib._internal.utils.numpy_utils.type import infer_dtype
 
 __all__ = [
     "BOOL_SENTINEL",
