+++ python/michelangelo/lib/_internal/utils/numpy_utils/sentinel.py
