+++ python/michelangelo/lib/_internal/utils/numpy_utils/type.py
